+++ lib/an_viewmodel.dart
@@ -1,5 +1,6 @@
 library an_viewmodel;
 
+export 'src/ext/context_ext.dart';
 export 'src/ext/value_notifier_ext.dart';
 export 'src/ext/vm_ext.dart';
 export 'src/ext/vm_ext_advanced.dart';